modified folder structure and separated out the superset-core files
