+++ superset-frontend/packages/superset-plugin-chart-immersatable/src/plugin/index.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { t, ChartMetadata, ChartPlugin } from '@superset-ui/core';
-import buildQuery from './buildQuery';
+import buildQuery from './superset-core/buildQuery';
 import controlPanel from './controlPanel';
 import transformProps from './transformProps';
 import thumbnail from '../images/thumbnail.png';
